refactor: extract shared search handler to eliminate route duplication
